pyproject.toml
[project]
name = "bioinformatics-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcbio-gff>=0.7.1",
    "biopython>=1.86",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
]
